requirements.txt
telethon==1.36.0
# openai==1.54.4  # Optional - now handled by n8n
aiohttp==3.10.10
python-dotenv==1.0.1
httpx==0.27.2
jsonschema==4.21.1
cryptography==43.0.3 